+++ tests/integration/conversations/v1/conversation/test_message.py
@@ -44,7 +44,18 @@ class MessageTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "index": 0,
-                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "delivery": {
+                    "total": 2,
+                    "sent": "all",
+                    "delivered": "some",
+                    "read": "some",
+                    "failed": "none",
+                    "undelivered": "none"
+                },
+                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                }
             }
             '''
         ))
@@ -77,7 +88,18 @@ class MessageTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "index": 0,
-                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "delivery": {
+                    "total": 2,
+                    "sent": "all",
+                    "delivered": "some",
+                    "read": "some",
+                    "failed": "none",
+                    "undelivered": "none"
+                },
+                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                }
             }
             '''
         ))
@@ -117,7 +139,18 @@ class MessageTestCase(IntegrationTestCase):
                 "date_created": "2015-12-16T22:18:37Z",
                 "date_updated": "2015-12-16T22:18:38Z",
                 "index": 0,
-                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "delivery": {
+                    "total": 2,
+                    "sent": "all",
+                    "delivered": "some",
+                    "read": "some",
+                    "failed": "none",
+                    "undelivered": "none"
+                },
+                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                }
             }
             '''
         ))
@@ -180,7 +213,18 @@ class MessageTestCase(IntegrationTestCase):
                 "date_created": "2016-03-24T20:37:57Z",
                 "date_updated": "2016-03-24T20:37:57Z",
                 "index": 0,
-                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "delivery": {
+                    "total": 2,
+                    "sent": "all",
+                    "delivered": "some",
+                    "read": "some",
+                    "failed": "none",
+                    "undelivered": "none"
+                },
+                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "links": {
+                    "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                }
             }
             '''
         ))
@@ -229,7 +273,18 @@ class MessageTestCase(IntegrationTestCase):
                         "date_created": "2016-03-24T20:37:57Z",
                         "date_updated": "2016-03-24T20:37:57Z",
                         "index": 0,
-                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "delivery": {
+                            "total": 2,
+                            "sent": "all",
+                            "delivered": "some",
+                            "read": "some",
+                            "failed": "none",
+                            "undelivered": "none"
+                        },
+                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "links": {
+                            "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                        }
                     },
                     {
                         "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
@@ -243,7 +298,18 @@ class MessageTestCase(IntegrationTestCase):
                         "date_created": "2016-03-24T20:38:21Z",
                         "date_updated": "2016-03-24T20:38:21Z",
                         "index": 0,
-                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "delivery": {
+                            "total": 2,
+                            "sent": "all",
+                            "delivered": "some",
+                            "read": "some",
+                            "failed": "none",
+                            "undelivered": "none"
+                        },
+                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "links": {
+                            "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                        }
                     },
                     {
                         "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
@@ -264,7 +330,18 @@ class MessageTestCase(IntegrationTestCase):
                         "date_created": "2016-03-24T20:38:21Z",
                         "date_updated": "2016-03-24T20:38:21Z",
                         "index": 0,
-                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "delivery": {
+                            "total": 2,
+                            "sent": "all",
+                            "delivered": "some",
+                            "read": "some",
+                            "failed": "none",
+                            "undelivered": "none"
+                        },
+                        "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "links": {
+                            "delivery_receipts": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Receipts"
+                        }
                     }
                 ]
             }