+++ tests/integration/conversations/v1/conversation/test_participant.py
@@ -149,6 +149,33 @@ class ParticipantTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
+    def test_update_gmms_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "conversation_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "sid": "MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "identity": "id",
+                "attributes": "{ \\"role\\": \\"driver\\" }",
+                "messaging_binding": {
+                    "type": "sms",
+                    "projected_address": "+15017122661"
+                },
+                "role_sid": "RLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "date_created": "2015-12-16T22:18:37Z",
+                "date_updated": "2015-12-16T22:18:38Z",
+                "url": "https://conversations.twilio.com/v1/Conversations/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/MBaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+            }
+            '''
+        ))
+
+        actual = self.client.conversations.v1.conversations("CHXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX") \
+                                             .participants("MBXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX").update()
+
+        self.assertIsNotNone(actual)
+
     def test_delete_request(self):
         self.holodeck.mock(Response(500, ''))