+++ tests/integration/flex_api/v1/test_configuration.py
@@ -136,22 +136,6 @@ class ConfigurationTestCase(IntegrationTestCase):
                         "config": "{\\"callback\\":\\"twilio.com/cb\\",\\"allowed_methods\\":[\\"GET\\",\\"POST\\"]}"
                     }
                 ],
-                "wfm_integrations": [
-                    {
-                        "name": "teleopti",
-                        "type": "flex_wfm",
-                        "active": true,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}",
-                        "logo": "logo1",
-                        "author": "somebody1"
-                    },
-                    {
-                        "name": "teleopti-stage",
-                        "type": "flex_wfm",
-                        "active": false,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}"
-                    }
-                ],
                 "outbound_call_flows": {
                     "default": {
                         "caller_id": "+12345",
@@ -329,22 +313,6 @@ class ConfigurationTestCase(IntegrationTestCase):
                         "config": "{\\"callback\\":\\"twilio.com/cb\\",\\"allowed_methods\\":[\\"GET\\",\\"POST\\"]}"
                     }
                 ],
-                "wfm_integrations": [
-                    {
-                        "name": "teleopti",
-                        "type": "flex_wfm",
-                        "active": true,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}",
-                        "logo": "logo1",
-                        "author": "somebody1"
-                    },
-                    {
-                        "name": "teleopti-stage",
-                        "type": "flex_wfm",
-                        "active": false,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}"
-                    }
-                ],
                 "outbound_call_flows": {
                     "default": {
                         "caller_id": "+12345",
@@ -520,22 +488,6 @@ class ConfigurationTestCase(IntegrationTestCase):
                         "type": "http",
                         "active": false,
                         "config": "{\\"callback\\":\\"twilio.com/cb\\",\\"allowed_methods\\":[\\"GET\\",\\"POST\\"]}"
-                    }
-                ],
-                "wfm_integrations": [
-                    {
-                        "name": "teleopti",
-                        "type": "flex_wfm",
-                        "active": true,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}",
-                        "logo": "logo1",
-                        "author": "somebody1"
-                    },
-                    {
-                        "name": "teleopti-stage",
-                        "type": "flex_wfm",
-                        "active": false,
-                        "config": "{\\"rta_callback_url\\":\\"https://twilio.com\\",\\"rta_feed_enabled\\": true, \\"historical_feed_enabled\\": true}"
                     }
                 ],
                 "outbound_call_flows": {